+++ composeApp/src/main/kotlin/com/zakir/vestra/MainActivity.kt
@@ -66,6 +66,7 @@ class MainActivity : ComponentActivity() {
                         runDiagnostics = app.runDiagnostics,
                         localJobStore = app.localJobStore,
                         chatRepository = app.chatRepository,
+                        memoryRepository = app.memoryRepository,
                         deviceRamMb = app.deviceProbe.totalRamMb(),
                         freeCloudDiscovery = app.freeCloudDiscovery,
                         humanParsing = app.humanParsing,